Added comments and naming convention
