+++ QAMGenFS2019/main.c
@@ -1,8 +1,8 @@
 /*
- * QAMGenFS2019.c
+ * FreeRTOS10_Template2.c
  *
  * Created: 20.03.2018 18:32:07
- * Author : chaos, Michael Meier
+ * Author : chaos
  */ 
 
 //#include <avr/io.h>
@@ -21,19 +21,18 @@
 
 #include "mem_check.h"
 
-#include "dma.h"
+#include "outputManagement.h"
 #include "init.h"
 #include "utils.h"
 #include "errorHandler.h"
 #include "NHD0420Driver.h"
 
-#include "tasks.h"
+#include "qamSendByte.h"
+
 
 extern void vApplicationIdleHook( void );
-void vLedBlink(void *pvParameters);
 
-TaskHandle_t ledTask;
-TaskHandle_t TaskDMAHandler;
+TaskHandle_t xTaskDMAHandler;
 
 void vApplicationIdleHook( void )
 {	
@@ -42,42 +41,16 @@ void vApplicationIdleHook( void )
 
 int main(void)
 {
-    resetReason_t reason = getResetReason();
-
 	vInitClock();
 	vInitDisplay();
 	
-	xTaskCreate( vLedBlink, (const char *) "ledBlink", configMINIMAL_STACK_SIZE+10, NULL, 1, &ledTask);
-	xTaskCreate( vTask_DMAHandler, (const char *) "dmaHandler", configMINIMAL_STACK_SIZE + 100, NULL, 1, &TaskDMAHandler);
-
+	xTaskCreate( vTask_DMAHandler, (const char *) "dmaHandler", configMINIMAL_STACK_SIZE + 300, NULL, 1, &xTaskDMAHandler);
+	
 	vInitDAC();
 	vInitDMATimer();
 	vInitDMA();
-
-
-	vDisplayClear();
-	vDisplayWriteStringAtPos(0,0,"FreeRTOS 10.0.1");
-	vDisplayWriteStringAtPos(1,0,"EDUBoard 1.0");
-	vDisplayWriteStringAtPos(2,0,"Template");
-	vDisplayWriteStringAtPos(3,0,"ResetReason: %d", reason);
 	vTaskStartScheduler();
 	return 0;
 }
 
-void vLedBlink(void *pvParameters) {
-	(void) pvParameters;
-	PORTF.DIRSET = PIN0_bm; /*LED1*/
-	PORTF.OUT = 0x01;
-	for(;;) {
-// 		uint32_t stack = get_mem_unused();
-// 		uint32_t heap = xPortGetFreeHeapSize();
-// 		uint32_t taskStack = uxTaskGetStackHighWaterMark(ledTask);
-// 		vDisplayClear();
-// 		vDisplayWriteStringAtPos(0,0,"Stack: %d", stack);
-// 		vDisplayWriteStringAtPos(1,0,"Heap: %d", heap);
-// 		vDisplayWriteStringAtPos(2,0,"TaskStack: %d", taskStack);
-// 		vDisplayWriteStringAtPos(3,0,"FreeSpace: %d", stack+heap);
-		PORTF.OUTTGL = 0x01;				
-		vTaskDelay(100 / portTICK_RATE_MS);
-	}
-}
+
